requirements.txt
progressbar2==3.12.0
acora==2.0
lxml==3.7.3
intervaltree==2.1.0
mock==2.0.0
scikit-learn==0.18.1
numpy==1.12.0
scipy==0.18.1
txttk==0.10.0
